requirements.txt
pyTelegramBotAPI==4.7.1
python-dotenv==0.21.0
requests==2.28.1
